Implemented usePatches custom hook to abstract Patch related functionality from New Asset component
