+++ src/components/new-asset-form/new-asset.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useReducer, useState } from "react";
+import React, { useState } from "react";
 import { Link as RouterLink } from "react-router-dom";
 
 import { Field, Form, Formik } from "formik";
@@ -18,49 +18,10 @@ import { NewPropertyFormData, newPropertySchema } from "./schema";
 import { renderAreaOfficeNamesOptions } from "./utils/area-office-names";
 import { renderManagingOrganisationOptions } from "./utils/managing-organisations";
 
-import { Center, Spinner, axiosInstance } from "@mtfh/common";
+import { Center, Spinner } from "@mtfh/common";
 import { createAsset } from "@mtfh/common/lib/api/asset/v1";
 import { CreateNewAssetRequest } from "@mtfh/common/lib/api/asset/v1/types";
-import { Patch } from "@mtfh/common/lib/api/patch/v1/types";
-import PropertyPatch from "../../utils/patch";
-
-const initialPatchesState = {
-  patches: [new PropertyPatch(1)],
-}
-
-function reducer(state: any, action: any) {
-  switch (action.type) {
-    case 'add_patch':
-      return {
-        patches: [...state.patches, action.payload],
-      }
-    case 'remove_patch':
-      return {
-        patches: state.patches.filter(
-          (patch: any) => patch.id !== action.payload.id
-        ),
-      }
-    case 'patch_edit': {
-      const patchIndex = state.patches.findIndex(
-        (patch: any) => patch.id == action.payload.patchId
-      )
-
-      state.patches[patchIndex].value = action.payload.targetValue
-      return { patches: state.patches }
-    }
-    default:
-      return state
-  }
-}
-
-const getAllPatchesAndAreas = async (): Promise<Array<Patch>> => {
-  return new Promise<Array<Patch>>((resolve, reject) => {
-    axiosInstance
-      .get<Array<Patch>>("https://2eqzdakhz2.execute-api.eu-west-2.amazonaws.com/development/api/v1/patch/all")
-      .then((res) => resolve(res.data))
-      .catch((error) => reject(error));
-  });
-};
+import { usePatches } from "./utils/usePatches";
 
 export interface Props {
   setShowSuccess: (value: boolean) => void;
@@ -78,16 +39,7 @@ export const NewAsset = ({
   setNewProperty,
 }: Props) => {
   const [loading, setLoading] = useState<boolean>(false);
-
-  // This state is used to manage the Patch field(s) in the form
-  const [patchesState, dispatch] = useReducer(reducer, initialPatchesState);
-
-  const [patchesAndAreasData, setPatchesAndAreasData] = useState<Patch[]>([]);
-
-  useEffect(() => {
-    getAllPatchesAndAreas().then(res => setPatchesAndAreasData(res));
-  }, [])
-
+  const { getFullPatchData, patchesState, renderPatchFormField} = usePatches();
 
   const renderAssetTypeOptions = (): JSX.Element[] => {
     return Object.keys(AssetType).map((key, index) => (
@@ -96,102 +48,6 @@ export const NewAsset = ({
       </option>
     ));
   };
-
-  const renderPatchOptions = (): JSX.Element[] | undefined => {
-    if (patchesAndAreasData) {
-      return patchesAndAreasData.map(({ id, name }) => (
-        <option key={id} value={id}>
-          {name}
-        </option>
-      ));
-    }
-  };
-
-  const generateNewPropertyPatchId = () => {
-    const assignedIds = patchesState.patches.map((patch: any) => patch.id)
-
-    // If there are no patches, there will be no assigned Ids,
-    // so we start with ID 1, otherwise we pick the next higher/available one.
-    return assignedIds.length == 0 ? 1 : Math.max(...assignedIds) + 1
-  }
-
-  const renderPropertyPatches = () => {
-    if (patchesAndAreasData.length) {
-      const patches = patchesState.patches.map((patch: any) => {
-        return (
-          <>
-            <div id="patch" key={patch.id}>
-              <Field
-                as="select"
-                id="patches"
-                name="patches"
-                className="govuk-input lbh-input"
-                data-testid="patches"
-                value={patch.value}
-                onChange={(e: any) =>
-                  handlePatchEdit(e, patch.id)
-                }
-              >
-                <option disabled value="">
-                  {" "}
-                  -- Select an option --{" "}
-                </option>
-                {renderPatchOptions()}
-              </Field>
-              <button
-                className="lbh-link"
-                role="button"
-                onClick={(e) => handleRemovePatch(e, patch)}
-                data-testid="patch-remove-link"
-                id="patch-remove-link"
-              >
-                Remove patch
-              </button>
-            </div>
-          </>
-        )
-      })
-      return patches;
-    } else {
-      return (
-        <div>
-          <Spinner />
-        </div>
-      )
-    }
-  }
-
-  const handleAddNewPatch = (e: any) => {
-    e.preventDefault()
-    dispatch({
-      type: 'add_patch',
-      payload: new PropertyPatch(generateNewPropertyPatchId()),
-    })
-  }
-
-  const handleRemovePatch = (e: any, patch: any) => {
-    e.preventDefault()
-    dispatch({ type: 'remove_patch', payload: patch })
-  }
-
-  const handlePatchEdit = (e: any, patchId: any) => {
-    dispatch({
-      type: 'patch_edit',
-      payload: {
-        targetValue: e.target.value,
-        patchId: patchId,
-      },
-    })
-  }
-
-  const getFullPatchData = (patchesState: any) => {
-    // Get patches GUIDs from patchesState
-    const patchesGuids = patchesState.patches.map((patch: PropertyPatch) => patch.value);
-    console.log("patchesGUIDS array", patchesGuids);
-
-    // Return full patch objects with the above GUIDs in patchesAndAreasData
-    return patchesAndAreasData.filter((patchObject: Patch) => patchesGuids.includes(patchObject.id))
-  }
 
   const handleSubmit = async (values: NewPropertyFormData) => {
     setShowSuccess(false);
@@ -730,33 +586,7 @@ export const NewAsset = ({
                   </div>
                 </fieldset>
               </div>
-              <label className="govuk-label lbh-label" htmlFor="patches">
-                Patches
-              </label>
-              <div id="property-patches-container">
-                {renderPropertyPatches()}
-              </div>
-              <div>
-                {patchesState.patches.length == 0 ? (
-                  <a
-                    className="lbh-link"
-                    href="#"
-                    onClick={(e) => handleAddNewPatch(e)}
-                    data-testid="add-patch-link"
-                  >
-                    Add a patch
-                  </a>
-                ) : (
-                  <a
-                    className="lbh-link"
-                    href="#"
-                    onClick={(e) => handleAddNewPatch(e)}
-                    data-testid="add-patch-link"
-                  >
-                    Add another patch
-                  </a>
-                )}
-              </div>
+              {renderPatchFormField()}
               <h2 className="lbh-heading-h2">Asset details</h2>
               {assetIsOfDwellingType(values.assetType) && (
                 <>
